requirements.txt
apt-xapian-index==0.45
apturl===0.4.1ubuntu4
attrs==16.3.0
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
BeautifulSoup==3.2.1
boto==2.41.0
bz2file==0.98
certifi==2016.9.26
chardet==2.0.1
colorama==0.2.5
command-not-found==0.3
configglue==1.1.2
configobj==4.7.2
configparser==3.5.0
constantly==15.1.0
coverage==4.2
cssselect==1.0.0
cycler==0.10.0
dask==0.11.0
debtagshw==0.1
decorator==4.0.10
defer==1.0.6
dirspec==13.10
dnspython==1.11.1
duplicity==0.6.23
entrypoints==0.2.2
feedparser==5.1.3
functools32==3.2.3.post2
gensim==0.13.1
gevent==1.0
greenlet==0.4.2
h5py==2.6.0
html5lib==0.999
httplib2==0.8
imbalanced-learn==0.2.1
incremental==16.10.1
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
jieba==0.38
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.0
Keras==1.0.5
Lasagne==0.1
lockfile==0.8
lxml==3.3.3
Mako==0.9.1
MarkupSafe==0.18
matplotlib==1.5.1
maxent===version-devel
mistune==0.7.3
mysql==0.0.1
mysql-connector==2.1.3
MySQL-python==1.2.5
nbconvert==4.2.0
nbformat==4.1.0
nemo-emblems==0.0.1
networkx==1.11
nltk==3.2.1
notebook==4.2.3
numpy==1.12.1
oauthlib==0.6.1
oneconf==0.3.7.14.4.1
OpenCC==0.2
PAM==0.4.2
pandas==0.18.1
pandoc==1.0.0b2
paramiko==1.10.1
parsel==1.1.0
pathlib2==2.1.0
pexpect==3.1
pickleshare==0.7.4
Pillow==2.3.0
piston-mini-client==0.7.5
ply==3.9
prompt-toolkit==1.0.7
protobuf==2.5.0
ptyprocess==0.5.1
py==1.4.31
py2neo==3.1.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
PyDispatcher==2.0.5
Pygments==2.1.3
pygobject==3.12.0
pyinotify==0.9.4
pyOpenSSL==0.13
pyparsing==2.1.5
pyserial==2.6
pysmbc==1.0.14.1
PyStanfordDependencies==0.3.1
pytest==3.0.1
python-apt===0.9.3.5ubuntu1
python-dateutil==2.5.3
python-debian===0.1.21-nmu2ubuntu2
python-igraph==0.7.1.post6
python-louvain==0.5
pytz==2016.4
pyxdg==0.25
PyYAML==3.11
pyzmq==16.0.0
qtconsole==4.2.1
queuelib==1.4.2
reportlab==3.0
requests==2.2.1
scikit-image==0.12.3
scikit-learn==0.18.1
scipy==0.19.0
Scrapy==1.3.0
service-identity==16.0.0
sessioninstaller==0.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
smart-open==1.3.3
spark-sklearn==0.2.0
stop-words==2015.2.23.1
system-service==0.1.6
terminado==0.6
Theano==0.8.2
toolz==0.8.0
toree==0.1.0.dev8
tornado==4.4.2
traitlets==4.2.2
Twisted==16.6.0
Twisted-Core==13.2.0
Twisted-Names==13.2.0
Twisted-Web==13.2.0
urllib3==1.7.1
uTidylib==0.2
w3lib==1.16.0
wcwidth==0.1.7
web.py==0.37
widgetsnbextension==1.2.6
zope.interface==4.0.5
